move handle_if_not_exist
