+++ eva/executor/create_executor.py
@@ -12,9 +12,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from eva.binder.binder_utils import handle_if_not_exists
 from eva.catalog.catalog_manager import CatalogManager
 from eva.executor.abstract_executor import AbstractExecutor
+from eva.executor.executor_utils import handle_if_not_exists
 from eva.planner.create_plan import CreatePlan
 from eva.storage.storage_engine import StorageEngine
