+++ eva/executor/executor_utils.py
@@ -14,8 +14,11 @@
 # limitations under the License.
 from typing import List
 
+from eva.catalog.catalog_manager import CatalogManager
 from eva.expression.abstract_expression import AbstractExpression
 from eva.models.storage.batch import Batch
+from eva.parser.table_ref import TableInfo
+from eva.utils.logging_manager import logger
 
 
 class ExecutorError(Exception):
@@ -35,3 +38,18 @@ def apply_predicate(batch: Batch, predicate: AbstractExpression) -> Batch:
         batch.drop_zero(outcomes)
         batch.reset_index()
     return batch
+
+
+def handle_if_not_exists(table_info: TableInfo, if_not_exist=False):
+    if CatalogManager().check_table_exists(
+        table_info.database_name, table_info.table_name
+    ):
+        err_msg = "Table: {} already exists".format(table_info)
+        if if_not_exist:
+            logger.warn(err_msg)
+            return True
+        else:
+            logger.error(err_msg)
+            raise ExecutorError(err_msg)
+    else:
+        return False
